make it more compact

+++ src/analysis/function_analysis.cpp
@@ -266,7 +266,7 @@ private:
                     assert(name->vreg == -1);
                     // skip
                 } else {
-                    RELEASE_ASSERT(0, "%d", name->lookup_type);
+                    RELEASE_ASSERT(0, "%u", (unsigned int)name->lookup_type);
                 }
                 break;
             }
